+++ BTDB/ODBLayer/ODBDictionary.cs
@@ -83,7 +83,7 @@ namespace BTDB.ODBLayer
             if (config.FreeContent == null)
             {
                 var method = ILBuilder.Instance
-                .NewMethod<Action<IInternalObjectDBTransaction, AbstractBufferedReader, IList<ulong>, IList<ulong>>>(
+                .NewMethod<Action<IInternalObjectDBTransaction, AbstractBufferedReader, IList<ulong>>>(
                     $"IDictFinder_Cfg_{cfgId}");
                 var ilGenerator = method.Generator;
 
@@ -92,8 +92,7 @@ namespace BTDB.ODBLayer
                     .Ldarg(0)
                     .Ldarg(1)
                     .Ldarg(2)
-                    .Ldarg(3)
-                    .Newobj(() => new DBReaderWithFreeInfoCtx(null, null, null, null))
+                    .Newobj(() => new DBReaderWithFreeInfoCtx(null, null, null))
                     .Stloc(readerLoc);
 
                 Action<IILGen> readerOrCtx;
@@ -106,7 +105,7 @@ namespace BTDB.ODBLayer
                 config.FreeContent = method.Create();
             }
 
-            var findIDictAction = (Action<IInternalObjectDBTransaction, AbstractBufferedReader, IList<ulong>, IList <ulong>>)config.FreeContent;
+            var findIDictAction = (Action<IInternalObjectDBTransaction, AbstractBufferedReader, IList<ulong>>)config.FreeContent;
 
             long prevProtectionCounter = 0;
             long pos = 0;
@@ -134,7 +133,7 @@ namespace BTDB.ODBLayer
                 prevProtectionCounter = _keyValueTrProtector.ProtectionCounter;
                 var valueBytes = _keyValueTr.GetValueAsByteArray();
                 var valueReader = new ByteArrayReader(valueBytes);
-                findIDictAction(ctx.GetTransaction(), valueReader, ctx.DictIds, ctx.Oids);
+                findIDictAction(ctx.GetTransaction(), valueReader, ctx.DictIds);
 
                 pos++;
             }