+++ BTDB/ODBLayer/ObjectDBTransaction.cs
@@ -82,7 +82,7 @@ namespace BTDB.ODBLayer
             if (tableInfo == null) throw new BTDBException($"Unknown TypeId {tableId} of inline object");
             var freeContentTuple = tableInfo.GetFreeContent(tableVersion);
             var readerWithFree = (DBReaderWithFreeInfoCtx)readerCtx;
-            freeContentTuple.Item2(this, null, reader, readerWithFree.DictIds, readerWithFree.Oids);
+            freeContentTuple.Item2(this, null, reader, readerWithFree.DictIds);
         }
 
         public void WriteInlineObject(object @object, IWriterCtx writerCtx)