+++ BTDB/ODBLayer/RelationDBManipulator.cs
@@ -45,13 +45,14 @@ namespace BTDB.ODBLayer
 
         static ByteBuffer EmptyBuffer = ByteBuffer.NewEmpty();
 
-        IRelationModificationCounter _modificationCounter;
+        readonly IRelationModificationCounter _modificationCounter;
 
         public RelationDBManipulator(IObjectDBTransaction transation, RelationInfo relationInfo)
         {
             _transaction = (IInternalObjectDBTransaction)transation;
             _relationInfo = relationInfo;
             _modificationCounter = _transaction.GetRelationModificationCounter(relationInfo.Id);
+            _hasSecondaryIndexes = _relationInfo.ClientRelationVersionInfo.HasSecondaryIndexes;
         }
 
         public IRelationModificationCounter ModificationCounter => _modificationCounter;
@@ -78,7 +79,7 @@ namespace BTDB.ODBLayer
             _transaction.KeyValueDBTransaction.SetKeyPrefix(ObjectDB.AllRelationsPKPrefix);
         }
 
-        bool HasSecondaryIndexes => _relationInfo.ClientRelationVersionInfo.HasSecondaryIndexes;
+        readonly bool _hasSecondaryIndexes;
 
         public bool Insert(T obj)
         {
@@ -91,9 +92,10 @@ namespace BTDB.ODBLayer
 
             if (_transaction.KeyValueDBTransaction.Find(keyBytes) == FindResult.Exact)
                 return false;
+
             _transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes);
 
-            if (HasSecondaryIndexes)
+            if (_hasSecondaryIndexes)
                 AddIntoSecondaryIndexes(obj);
 
             _modificationCounter.MarkModification();
@@ -114,7 +116,7 @@ namespace BTDB.ODBLayer
 
                 _transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes);
 
-                if (HasSecondaryIndexes)
+                if (_hasSecondaryIndexes)
                     UpdateSecondaryIndexes(obj, keyBytes, oldValueBytes);
 
                 FreeContentInUpdate(oldValueBytes, valueBytes);
@@ -122,8 +124,45 @@ namespace BTDB.ODBLayer
             }
 
             _transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes);
-            if (HasSecondaryIndexes)
+            if (_hasSecondaryIndexes)
                 AddIntoSecondaryIndexes(obj);
+            _modificationCounter.MarkModification();
+            return true;
+        }
+
+        public bool ShallowUpsert(T obj)
+        {
+            Debug.Assert(typeof(T) == obj.GetType(), AssertNotDerivedTypesMsg);
+
+            var keyBytes = KeyBytes(obj);
+            var valueBytes = ValueBytes(obj);
+
+            StartWorkingWithPK();
+
+            if (_hasSecondaryIndexes)
+            {
+                if (_transaction.KeyValueDBTransaction.Find(keyBytes) == FindResult.Exact)
+                {
+                    var oldValueBytes = _transaction.KeyValueDBTransaction.GetValue();
+
+                    _transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes);
+
+                    UpdateSecondaryIndexes(obj, keyBytes, oldValueBytes);
+
+                    return false;
+                }
+
+                _transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes);
+                AddIntoSecondaryIndexes(obj);
+            }
+            else
+            {
+                if (!_transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes))
+                {
+                    return false;
+                }
+            }
+
             _modificationCounter.MarkModification();
             return true;
         }
@@ -143,10 +182,32 @@ namespace BTDB.ODBLayer
             var oldValueBytes = _transaction.KeyValueDBTransaction.GetValue();
             _transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes);
 
-            if (HasSecondaryIndexes)
+            if (_hasSecondaryIndexes)
                 UpdateSecondaryIndexes(obj, keyBytes, oldValueBytes);
 
             FreeContentInUpdate(oldValueBytes, valueBytes);
+        }
+
+        public void ShallowUpdate(T obj)
+        {
+            Debug.Assert(typeof(T) == obj.GetType(), AssertNotDerivedTypesMsg);
+
+            var keyBytes = KeyBytes(obj);
+            var valueBytes = ValueBytes(obj);
+
+            StartWorkingWithPK();
+
+            if (_transaction.KeyValueDBTransaction.Find(keyBytes) != FindResult.Exact)
+                throw new BTDBException("Not found record to update.");
+
+            var oldValueBytes = _hasSecondaryIndexes
+                ? _transaction.KeyValueDBTransaction.GetValue()
+                : ByteBuffer.NewEmpty();
+
+            _transaction.KeyValueDBTransaction.CreateOrUpdateKeyValue(keyBytes, valueBytes);
+
+            if (_hasSecondaryIndexes)
+                UpdateSecondaryIndexes(obj, keyBytes, oldValueBytes);
         }
 
         public bool Contains(ByteBuffer keyBytes)
@@ -210,7 +271,7 @@ namespace BTDB.ODBLayer
             var valueBytes = _transaction.KeyValueDBTransaction.GetValue();
             _transaction.KeyValueDBTransaction.EraseCurrent();
 
-            if (HasSecondaryIndexes)
+            if (_hasSecondaryIndexes)
                 RemoveSecondaryIndexes(keyBytes, valueBytes);
 
             _relationInfo.FreeContent(_transaction, valueBytes);
@@ -236,7 +297,7 @@ namespace BTDB.ODBLayer
 
                 var valueBytes = _transaction.KeyValueDBTransaction.GetValue();
 
-                if (HasSecondaryIndexes)
+                if (_hasSecondaryIndexes)
                     RemoveSecondaryIndexes(key, valueBytes);
 
                 if (_relationInfo.NeedImplementFreeContent())
@@ -265,7 +326,7 @@ namespace BTDB.ODBLayer
 
         public int RemoveByKeyPrefixWithoutIterate(ByteBuffer keyBytesPrefix)
         {
-            if (HasSecondaryIndexes)
+            if (_hasSecondaryIndexes)
             {
                 //keyBytePrefix contains [Index Relation, Primary key prefix] we need
                 //                       [Index Relation, Secondary Key Index, Primary key prefix]